MS-574 Milvus configuration refactor

See merge request megasearch/milvus!610

Former-commit-id: 9d0f05589f4344bb029c2f65470d9b567999ae3f

+++ cpp/src/db/engine/ExecutionEngineImpl.h
@@ -101,7 +101,7 @@ protected:
     std::string location_;
 
     int32_t nlist_ = 0;
-    int64_t gpu_num_ = 0;
+    int32_t gpu_num_ = 0;
 };
